Be -> develop

See merge request s08-bigdata-recom-sub2/S08P22A505!116

+++ backend/billboard/src/main/java/com/ssafy/billboard/model/service/RoomServiceImpl.java
@@ -9,7 +9,6 @@ import com.ssafy.billboard.model.repository.ReplyRepository;
 import com.ssafy.billboard.model.repository.RoomRepository;
 import com.ssafy.billboard.model.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -25,9 +24,10 @@ public class RoomServiceImpl implements RoomService {
     private final EntryRepository entryRepository;
 
     @Override
-    public RoomDto.RoomInfo createRoom(RoomDto.RoomInput roomInput){
+    public long createRoom(RoomDto.RoomInput roomInput){
+        //없는 유저가 방장 > null
         if(!userRepository.existsByUserId(roomInput.getHostId()))
-            return null;
+            return -1;
         Room room = roomRepository.save(Room.builder()
                 .hostId(roomInput.getHostId())
                 .title(roomInput.getTitle())
@@ -37,17 +37,9 @@ public class RoomServiceImpl implements RoomService {
                 .lng(roomInput.getLng())
                 .date(roomInput.getDate())
                 .build());
-        return RoomDto.RoomInfo.builder()
-                .roomId(room.getRoomId())
-                .hostId(room.getHostId())
-                .title(room.getTitle())
-                .personCount(room.getEntries().size())
-                .personLimit(room.getPersonLimit())
-                .location(room.getLocation())
-                .lat(room.getLat())
-                .lng(room.getLng())
-                .date(room.getDate())
-                .build();
+        //방장은 자동 참여 처리
+        createEntry(room.getRoomId(), room.getHostId());
+        return room.getRoomId();
     }
 
     @Override
@@ -89,11 +81,13 @@ public class RoomServiceImpl implements RoomService {
                     .replies(getReplies(roomId))
                     .build();
         }
+        //없는 방아이디 > null
         return null;
     }
 
     @Override
     public boolean deleteRoom(long roomId){
+        //없는 방 아이디 > false
         if(!roomRepository.existsById(roomId))
             return false;
         roomRepository.deleteById(roomId);
@@ -102,16 +96,18 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public boolean updateRoom(long roomId, RoomDto.RoomUpdate roomUpdate){
+        //없는 방 아이디 > false
         if(!roomRepository.existsById(roomId))
             return false;
         Room room = roomRepository.findById(roomId).get();
-        room.update(roomUpdate);
+        room.updateRoom(roomUpdate);
         roomRepository.save(room);
         return true;
     }
 
     @Override
     public boolean createReply(long roomId, RoomDto.ReplyInput replyInput){
+        //없는 방 or 유저 > false;
         if(!roomRepository.existsById(roomId) || !userRepository.existsByUserId(replyInput.getUserId()))
             return false;
         replyRepository.save(Reply.builder()
@@ -124,6 +120,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public List<RoomDto.ReplyInfo> getReplies(long roomId){
+        //없는 방 아이디 > null
         if(!roomRepository.existsById(roomId))
             return null;
         List<Reply> repliesEntity = replyRepository.findAllByRoomId(roomId);
@@ -140,6 +137,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public boolean deleteReply(long replyId){
+        //없는 댓글 아이디 > false
         if(!replyRepository.existsById(replyId))
             return false;
         replyRepository.deleteById(replyId);
@@ -148,29 +146,37 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public int createEntry(long roomId, String userId){
+        //없는 방 or 없는 유저 > 0
         if(!roomRepository.existsById(roomId) || !userRepository.existsByUserId(userId))
             return 0;
+        //해당 방에 해당 유저가 이미 참여중 > -1
         if(entryRepository.existsByRoomAndUserId(roomRepository.findById(roomId).get(), userId))
             return -1;
+        Room room = roomRepository.findById(roomId).get();
         entryRepository.save(Entry.builder()
-                .room(roomRepository.findById(roomId).get())
+                .room(room)
                 .userId(userId)
                 .build());
         return 1;
     }
 
     @Override
-    public boolean deleteEntry(long roomId, String userId){
+    public int deleteEntry(long roomId, String userId){
+        //없는 방 or 유저 or 참여중이 아님 : 0
         if(!roomRepository.existsById(roomId)
                 || !userRepository.existsByUserId(userId)
                 || !entryRepository.existsByRoomAndUserId(roomRepository.findById(roomId).get(), userId))
-            return false;
+            return 0;
+        //방장이 참여 취소 : -1
+        if(roomRepository.findById(roomId).get().getHostId().equals(userId))
+            return -1;
         entryRepository.deleteByRoomAndUserId(roomRepository.findById(roomId).get(), userId);
-        return true;
+        return 1;
     }
 
     @Override
     public List<RoomDto.RoomReservationInfo> getRoomsByUserId(String userId){
+        //없는 유저 > null
         if(!userRepository.existsByUserId(userId))
             return null;
         List<Entry> entries = entryRepository.findAllByUserId(userId);
